move app bar to home screen
