+++ client_end_user/shared/src/commonMain/kotlin/di/screenModelsModule.kt
@@ -2,24 +2,22 @@ package di
 
 import org.koin.core.module.dsl.factoryOf
 import org.koin.dsl.module
-import presentation.preferredFood.PreferredFoodScreenModel
+import presentation.app.AppScreenModel
 import presentation.auth.login.LoginScreenModel
 import presentation.auth.signup.registration.RegistrationScreenModel
 import presentation.auth.signup.registrationSubmit.RegistrationSubmitScreenModel
 import presentation.cuisines.CuisinesScreenModel
 import presentation.home.HomeScreenModel
-import presentation.main.MainScreenModel
 import presentation.notification.NotificationScreenModel
 import presentation.orderHistory.OrderHistoryScreenModel
 import presentation.pickLanguage.PickLanguageScreenModel
+import presentation.preferredFood.PreferredFoodScreenModel
 import presentation.preferredMeal.PreferredScreenModel
 import presentation.preferredRide.PreferredRideScreenModel
-import presentation.app.AppScreenModel
 
 val screenModelsModule = module {
     factoryOf(::LoginScreenModel)
     factoryOf(::HomeScreenModel)
-    factoryOf(::MainScreenModel)
     factoryOf(::OrderHistoryScreenModel)
     factoryOf(::PreferredFoodScreenModel)
     factoryOf(::RegistrationScreenModel)
