+++ client_end_user/shared/src/commonMain/kotlin/presentation/home/HomeMappers.kt
@@ -2,6 +2,7 @@ package presentation.home
 
 import domain.entity.Offer
 import domain.entity.Restaurant
+import domain.entity.User
 
 fun Offer.toUiState(): OfferUiState {
     return OfferUiState(
@@ -19,3 +20,13 @@ fun Restaurant.toRestaurantUiState(): RestaurantUiState {
 }
 
 fun List<Restaurant>.toRestaurantUiState() = map { it.toRestaurantUiState() }
+
+
+
+fun User.toUIState() = UserUiState(
+    username = name,
+    wallet = walletValue,
+    currency = currency,
+    isLogin = true
+)
+
