+++ client_end_user/shared/src/commonMain/kotlin/presentation/home/HomeScreenInteractionListener.kt
@@ -11,4 +11,5 @@ interface HomeScreenInteractionListener : BaseInteractionListener {
     fun onClickOffersSlider(position : Int)
     fun onClickSearch()
     fun onClickOrderAgain(orderId : String)
+    fun onLoginClicked()
 }
