+++ client_end_user/shared/src/commonMain/kotlin/presentation/home/HomeScreenUiEffect.kt
@@ -9,4 +9,5 @@ sealed class HomeScreenUiEffect {
     data object ScrollDownToRecommendedRestaurants : HomeScreenUiEffect()
     data class NavigateToOfferItem(val offerId : String) : HomeScreenUiEffect()
     data object NavigateToSearch: HomeScreenUiEffect()
+    data object NavigateLoginScreen: HomeScreenUiEffect()
 }
