+++ client_end_user/shared/src/commonMain/kotlin/presentation/home/HomeScreenUiState.kt
@@ -10,6 +10,7 @@ data class HomeScreenUiState(
     val favoriteRestaurants: List<RestaurantUiState> = emptyList(),
     val inProgressWrapper: InProgressWrapper = InProgressWrapper(),
     val lastOrder : OrderUiState = OrderUiState(),
+    val user: UserUiState = UserUiState(),
 ) {
     val hasProgress: Boolean
         get() = inProgressWrapper.taxisOnTheWay.isNotEmpty() ||
@@ -38,3 +39,11 @@ data class RestaurantUiState(
     val rating: Double = 0.0,
     val priceLevel: PriceLevel = PriceLevel.LOW
 )
+
+data class UserUiState(
+    val username: String = "",
+    val isLogin: Boolean = false,
+    val wallet: Double = 0.0,
+    val currency: String = ""
+)
+
