+++ client_end_user/shared/src/commonMain/kotlin/presentation/main/MainContainer.kt
@@ -3,12 +3,9 @@ package presentation.main
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
@@ -21,76 +18,37 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.layout.positionInRoot
 import androidx.compose.ui.unit.dp
-import cafe.adriel.voyager.navigator.Navigator
+import cafe.adriel.voyager.core.screen.Screen
 import cafe.adriel.voyager.navigator.tab.CurrentTab
 import cafe.adriel.voyager.navigator.tab.LocalTabNavigator
 import cafe.adriel.voyager.navigator.tab.TabNavigator
-import com.beepbeep.designSystem.ui.composable.BpAppBar
-import com.beepbeep.designSystem.ui.composable.BpButton
 import com.beepbeep.designSystem.ui.composable.BpNavigationBar
 import com.beepbeep.designSystem.ui.composable.BpNavigationBarItem
 import com.beepbeep.designSystem.ui.theme.Theme
 import org.jetbrains.compose.resources.ExperimentalResourceApi
 import org.jetbrains.compose.resources.painterResource
-import presentation.base.BaseScreen
 import presentation.composable.exitinstion.drawTopIndicator
 import presentation.composable.exitinstion.toPx
-import presentation.auth.login.LoginScreen
-import resources.Resources
 import util.getNavigationBarPadding
 
 
-object MainContainer :
-    BaseScreen<MainScreenModel, MainUiState, MainScreenUiEffect, MainInteractionListener>() {
+object MainContainer : Screen {
 
     @Composable
     override fun Content() {
-        initScreen(getScreenModel())
-    }
-
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Composable
-    override fun onRender(state: MainUiState, listener: MainInteractionListener) {
         TabNavigator(HomeTab) {
             Scaffold(
-                topBar = {
-                    BpAppBar(
-                        title = if (state.isLogin) {
-                            Resources.strings.welcome + " ${state.username}"
-                        } else {
-                            Resources.strings.loginWelcomeMessage
-                        },
-                        actions = {
-                            if (state.isLogin) {
-                                Wallet(value = state.currency + state.wallet)
-                            } else {
-                                BpButton(
-                                    modifier = Modifier.heightIn(max = 32.dp).padding(end = 16.dp),
-                                    textPadding = PaddingValues(horizontal = 16.dp),
-                                    title = Resources.strings.login,
-                                    onClick = listener::onLoginClicked
-                                )
-                            }
-                        }
-                    )
-                },
                 bottomBar = {
                     val tabNavigator = LocalTabNavigator.current
                     val tabs = rememberTabsContainer()
                     BottomBar(tabs, tabNavigator)
                 },
                 content = {
-                    Column(Modifier.fillMaxSize().padding(it)) {
+                    Column(Modifier.fillMaxSize().padding(bottom = it.calculateBottomPadding())) {
                         CurrentTab()
                     }
                 }
             )
-        }
-    }
-
-    override fun onEffect(effect: MainScreenUiEffect, navigator: Navigator) {
-        when (effect) {
-            is MainScreenUiEffect.NavigateLoginScreen -> navigator.push(LoginScreen())
         }
     }
 
@@ -130,25 +88,6 @@ object MainContainer :
                     onClick = { tabNavigator.current = tabContainer.tab },
                 )
             }
-        }
-    }
-
-    @Composable
-    private fun Wallet(
-        modifier: Modifier = Modifier,
-        value: String
-    ) {
-        Column(modifier = modifier.padding(end = 16.dp)) {
-            Text(
-                Resources.strings.wallet,
-                style = Theme.typography.body,
-                color = Theme.colors.contentSecondary
-            )
-            Text(
-                value,
-                style = Theme.typography.title,
-                color = Theme.colors.primary
-            )
         }
     }
 }
