--- client_end_user/shared/src/commonMain/kotlin/presentation/main/MainInteractionListener.kt
@@ -1,7 +1,0 @@
-package presentation.main
-
-import presentation.base.BaseInteractionListener
-
-interface MainInteractionListener : BaseInteractionListener {
-    fun onLoginClicked()
-}
