--- client_end_user/shared/src/commonMain/kotlin/presentation/main/MainScreenUiEffect.kt
@@ -1,5 +1,0 @@
-package presentation.main
-
-sealed class MainScreenUiEffect {
-    object NavigateLoginScreen : MainScreenUiEffect()
-}
