--- client_end_user/shared/src/commonMain/kotlin/presentation/main/MainUiState.kt
@@ -1,19 +1,0 @@
-package presentation.main
-
-import domain.entity.User
-
-
-data class MainUiState(
-    val username: String = "",
-    val isLogin: Boolean = false,
-    val wallet: Double = 0.0,
-    val currency: String = ""
-)
-
-
-fun User.toUIState() = MainUiState(
-    username = name,
-    wallet = walletValue,
-    currency = currency,
-    isLogin = true
-)
